Make can_add_ulong_within return an int; it's a predicate.
The unsigned long doesn't make sense, and also it would cause conversion
warnings in ulong_add_capped.

+++ src/common.h
@@ -120,7 +120,7 @@ static inline int can_ulong_add(const unsigned long x,
         const unsigned long y) __ATTR_CONST;
 static inline unsigned long ulong_add(const unsigned long x,
         const unsigned long y) __ATTR_CONST;
-static inline unsigned long can_add_ulong_within(const unsigned long x,
+static inline int can_add_ulong_within(const unsigned long x,
         const unsigned long y, const unsigned long cap) __ATTR_CONST;
 static inline unsigned long ulong_add_capped(const unsigned long x,
         const unsigned long y, const unsigned long cap) __ATTR_CONST;
@@ -178,7 +178,7 @@ static inline unsigned long ulong_add(const unsigned long x,
 /*
  * Check whether two unsigned long can be added within a cap.
  */
-static inline unsigned long can_add_ulong_within(const unsigned long x,
+static inline int can_add_ulong_within(const unsigned long x,
         const unsigned long y, const unsigned long cap)
 {
     return y <= cap && x <= cap && x <= (cap - y);
